requirements.txt
fastapi
uvicorn
rembg[cpu]
pillow
python-multipart
jinja2
opencv-python-headless